# refactor(query): improve performance of aggregate function distinct