--- a/src/common/datavalues/src/macros.rs
+++ b/src/common/datavalues/src/macros.rs
@@ -283,29 +283,6 @@ macro_rules! with_match_primitive_types_error {
 }
 
 #[macro_export]
-macro_rules! with_match_primitive_types_error_2 {(
-        $key_type:expr, | $_:tt $T:ident | $($body:tt)*
-    ) => ({
-        macro_rules! __with_ty__ {( $_ $T:ident ) => ( $($body)* )}
-        type ordered_f32 = OrderedFloat<f32>;
-        type ordered_f64 = OrderedFloat<f64>;
-        match $key_type {
-            TypeID::Int8 => __with_ty__! { i8 },
-            TypeID::Int16 => __with_ty__! { i16 },
-            TypeID::Int32 => __with_ty__! { i32 },
-            TypeID::Int64 => __with_ty__! { i64 },
-            TypeID::UInt8 => __with_ty__! { u8 },
-            TypeID::UInt16 => __with_ty__! { u16 },
-            TypeID::UInt32 => __with_ty__! { u32 },
-            TypeID::UInt64 => __with_ty__! { u64 },
-            v => return Err(ErrorCode::BadDataValueType(
-                format!("Ops is not support on datatype: {:?}",v)
-            ))
-        }
-    })
-}
-
-#[macro_export]
 macro_rules! with_match_integer_type_id {
     ($key_type:expr, | $_:tt $T:ident | $body:tt,  $nbody:tt) => {{
         macro_rules! __with_ty__ {