Signed-off-by: kikkon <nian920@outlook.com>

I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

Summary about this PR

in debug mode:
before
```
mysql> select count(distinct number) from numbers_mt(10000000);
+------------------------+
| count(DISTINCT number) |
+------------------------+
|               10000000 |
+------------------------+
1 row in set (1 min 1.10 sec)
Read 10000000 rows, 76.29 MiB in 61.064 sec., 163.76 thousand rows/sec., 1.25 MiB/sec.
```

after
```
mysql> select count(distinct number) from numbers_mt(10000000);
+------------------------+
| count(DISTINCT number) |
+------------------------+
|               10000000 |
+------------------------+
1 row in set (22.38 sec)
Read 10000000 rows, 76.29 MiB in 22.338 sec., 447.66 thousand rows/sec., 3.42 MiB/sec.
```

## Changelog
improve performance of aggregate function distinct
## Related Issues
Related https://github.com/datafuselabs/databend/issues/6240