@sundy-li 2022-08-15
Update in release mode:

```
databend > select count(distinct number) from numbers(10000000);
+------------------------+
| count(DISTINCT number) |
+------------------------+
|               10000000 |
+------------------------+
1 row in set (8.171 sec)

databend > select count(distinct number) from numbers(10000000);
+------------------------+
| count(DISTINCT number) |
+------------------------+
|                     32 |
+------------------------+
1 row in set (0.272 sec)
```